Arreglo en ediciones y eliminaciones en sueldos, descuentos y asignaciones
Se arreglaron errores donde al querer editar un registro, el mismo indicaba que ya existía y no dejaba editarlo.
También se arregló el error de que, al querer eliminar un registro de un usuario, se eliminaban todos los registros del mismo usuario.
Se acomodó todo en diferentes carpetas para mantener una mejor organización

+++ BugTrackingSystem/BugTrackingSystem/Entidades/Sueldo.cs
@@ -6,7 +6,7 @@ namespace BugTrackingSystem.Entidades
     {
         public Usuario Usuario { get; set; }
         public DateTime Fecha { get; set; }
-        public float SueldoBruto { get; set; }
+        public decimal SueldoBruto { get; set; }
         public bool Borrado { get; set; }
     }
 }
